+++ Makefile
@@ -5,7 +5,8 @@ SRCS = \
 		project/src/main.c \
 		project/src/master_write.c \
 		project/src/transaction_write.c \
-		project/src/update_data.c
+		project/src/update_data.c \
+		project/src/file_reader.c
 
 .PHONY: all build rebuild check test memtest clean
