+++ project/include/file_reader.h
@@ -1,0 +1,9 @@
+#ifndef PROJECT_INCLUDE_FILE_READER_H_
+#define PROJECT_INCLUDE_FILE_READER_H_
+
+#include "utils.h"
+
+data_t *read_from_file_record(FILE *Ptr);
+data_t *read_from_file_transfer(FILE *Ptr);
+
+#endif  // PROJECT_INCLUDE_FILE_READER_H_
