+++ project/src/main.c
@@ -5,6 +5,7 @@
 #include "transaction_write.h"
 #include "master_write.h"
 #include "update_data.h"
+#include "file_reader.h"
 
 enum buf_sizes {
     FORMAT_STRING_MAX_SIZE = 110,
@@ -13,72 +14,6 @@ enum buf_sizes {
     ADDRESS = 30,
     TEL_NUMBER = 15,
 };
-
-int write_to_file_record(FILE *Ptr, data_t *data) {
-    if (!Ptr) {
-        return ERR_WRONG_POINTER;
-    }
-    return master_write(Ptr, data);
-}
-
-int write_to_file_transfer(FILE *Ptr, data_t *data) {
-    if (!Ptr) {
-        return ERR_WRONG_POINTER;
-    }
-    return transaction_write(Ptr, data);
-}
-
-data_t *read_from_file_record(FILE *Ptr) {
-    if (!Ptr) {
-        return NULL;
-    }
-    size_t data_size = sizeof(data_t);
-    data_t *data_array = malloc(data_size);
-    char format_string[FORMAT_STRING_MAX_SIZE];
-    snprintf(format_string, FORMAT_STRING_MAX_SIZE, "%%i%%%ds%%%ds%%%ds%%%ds%%lf%%lf%%lf\n",
-             NAME, SURNAME, ADDRESS, TEL_NUMBER);
-    unsigned int id = 0;
-    while (fscanf(Ptr, format_string, &data_array[id].number, data_array[id].name,
-                  data_array[id].surname, data_array[id].address,
-                  data_array[id].tel_number, &data_array[id].indebtedness,
-                  &data_array[id].credit_limit, &data_array[id].cash_payments) != -1) {
-        data_size += sizeof(data_t);
-        void *tmp = realloc(data_array, data_size);
-        if (tmp == NULL) {
-            free(data_array);
-            return NULL;
-        } else {
-            data_array = tmp;
-        }
-        ++id;
-    }
-    data_array[id].number = -1;
-    return data_array;
-}
-
-data_t *read_from_file_transfer(FILE *Ptr) {
-    if (!Ptr) {
-        return NULL;
-    }
-    size_t data_size = sizeof(data_t);
-    data_t *data_array = malloc(data_size);
-    char format_string[FORMAT_STRING_MAX_SIZE];
-    snprintf(format_string, FORMAT_STRING_MAX_SIZE, "%%i%%lf\n");
-    unsigned int id = 0;
-    while (fscanf(Ptr, format_string, &data_array[id].number, &data_array[id].cash_payments) != -1) {
-        data_size += sizeof(data_t);
-        void *tmp = realloc(data_array, data_size);
-        if (tmp == NULL) {
-            free(data_array);
-            return NULL;
-        } else {
-            data_array = tmp;
-        }
-        ++id;
-    }
-    data_array[id].number = -1;
-    return data_array;
-}
 
 data_t *input_data_record() {
     data_t *data = malloc(sizeof(data_t));
@@ -130,7 +65,7 @@ int main() {
                 } else {
                     record = input_data_record();
                     while (record != NULL) {
-                        if (write_to_file_record(Ptr, record) != 0) {
+                        if (master_write(Ptr, record) != 0) {
                             return ERR_WRONG_POINTER;
                         }
                         free(record);
@@ -146,7 +81,7 @@ int main() {
                 } else {
                     record = input_data_transfer();
                     while (record != NULL) {
-                        if (write_to_file_transfer(Ptr, record) != 0) {
+                        if (transaction_write(Ptr, record) != 0) {
                             return ERR_WRONG_POINTER;
                         }
                         free(record);
